colocate dark mode hooks
