+++ packages/app/navigation/RootStack.tsx
@@ -38,9 +38,9 @@ import { NotesDetailScreen } from '../features/top/NotesDetailScreen';
 import { NotesFolderScreen } from '../features/top/NotesFolderScreen';
 import PostScreen from '../features/top/PostScreen';
 import { UserProfileScreen } from '../features/top/UserProfileScreen';
+import { useIsDarkMode } from '../hooks/useDarkMode';
 import { useFeatureFlag } from '../lib/featureFlags';
 import { useTheme } from '../ui';
-import { useIsDarkMode } from '../ui/utils/colorUtils';
 import { GroupSettingsStack } from './GroupSettingsStack';
 import { TopLevelTabNavigator } from './TopLevelTabNavigator';
 import type { RootStackParamList } from './types';
